Update S10Q01_Capital_Letter.py
"""####Tile file:
Please solve the following 3 exercises
S10Q01
Get the name of the text file from the user. 
Check if all the sentences in that text file begin with a capital letter.
S10Q02
Write a Python script that takes a file name as its argument. 
In that file, find the line that has the maximum number of 
occurrences of the letter ‘e’
S10Q03
Write a Python script that takes a file name as its argument. 
Write the contents of this file to another file such that, 
   each sentence is stored on a new line.
"""