Create S17Q01_10_MostRepeatedWords.py
Output:
[suresh@localhost Python]$ python S17Q01_10_MostRepeatedWords.py demofile7.txt 
**********************************************************************
Top 10 most repeated words
the: 9
file: 6
a: 6
that: 5
of: 4
name: 3
Write: 3
text: 2
takes: 2
script: 2
Words that are not repeated
**********************************************************************
['all', 'sentence', 'is', 'sentences', 'Please', 'number', 'stored', 'occurrences', 'another', 'in', 'find', 'contents', 'if', 'that,', 'from', 'Get', 'to', '3', 'capital', 'new', 'has', 'begin', 'In', 'S10Q03', '\xe2\x80\x98e\xe2\x80\x99', 'S10Q01', 'exercises', 'such', 'with', 'file,', 'on', 'this', 'maximum', 'solve', 'S10Q02', 'each', 'following', 'Check', 'user']

Test file used:
[suresh@localhost Python]$ cat demofile7.txt
Please solve the following 3 exercises
S10Q01
Get the name of the text file from the user. 
Check if all the sentences in that text file begin with a capital letter.
S10Q02
Write a Python script that takes a file name as its argument. 
In that file, find the line that has the maximum number of 
occurrences of the letter ‘e’
S10Q03
Write a Python script that takes a file name as its argument. 
Write the contents of this file to another file such that, 
   each sentence is stored on a new line.